Organise tests and add pierce transit
