+++ onebusaway-gtfs/src/test/java/org/onebusaway/gtfs/serialization/GtfsReaderTest.java
@@ -50,7 +50,7 @@ import org.onebusaway.gtfs.services.GtfsMutableRelationalDao;
 import org.onebusaway.gtfs.services.GtfsRelationalDao;
 import org.onebusaway.gtfs.services.MockGtfs;
 
-public class GtfsReaderTest {
+public class GtfsReaderTest extends BaseGtfsTest {
 
   @Test
   public void testAllFields() throws IOException {
@@ -394,7 +394,7 @@ public class GtfsReaderTest {
   }
 
   @Test
-  public void testCaltrain() throws IOException, ParseException {
+  public void testCaltrain() throws IOException {
 
     File resourcePath = GtfsTestData.getCaltrainGtfs();
     String agencyId = "Caltrain";
@@ -555,7 +555,7 @@ public class GtfsReaderTest {
   }
 
   @Test
-  public void testBart() throws IOException, ParseException {
+  public void testBart() throws IOException {
 
     File resourcePath = GtfsTestData.getBartGtfs();
     String agencyId = "BART";
@@ -577,7 +577,7 @@ public class GtfsReaderTest {
   }
 
   @Test
-  public void testIntern() throws IOException, ParseException {
+  public void testIntern() throws IOException {
     File resourcePath;
     String agencyId;
     GtfsDao entityStore;
@@ -697,8 +697,7 @@ public class GtfsReaderTest {
   }
 
   @Test
-  public void testUtf8() throws IOException, ParseException,
-      InterruptedException {
+  public void testUtf8() throws IOException {
 
     MockGtfs mockGtfs = MockGtfs.create();
     mockGtfs.putDefaultStopTimes();
@@ -716,8 +715,7 @@ public class GtfsReaderTest {
   }
 
   @Test
-  public void testBom() throws IOException, ParseException,
-      InterruptedException {
+  public void testBom() throws IOException {
 
     MockGtfs mockGtfs = MockGtfs.create();
     mockGtfs.putDefaultStopTimes();
@@ -775,123 +773,6 @@ public class GtfsReaderTest {
     assertEquals(1, frequency.getExactTimes());
     assertEquals(300, frequency.getHeadwaySecs());
     assertSame(trip, frequency.getTrip());
-  }
-
-  @Test
-  public void turlockFaresV2() throws CsvEntityIOException, IOException {
-    String agencyId = "1642";
-    GtfsRelationalDao dao = processFeed(GtfsTestData.getTurlockFaresV2(),
-      agencyId, false);
-
-    Agency agency = dao.getAgencyForId(agencyId);
-    assertEquals(agencyId, agency.getId());
-    assertEquals("Turlock Transit", agency.getName());
-    assertEquals("http://www.turlocktransit.com/", agency.getUrl());
-    assertEquals("America/Los_Angeles", agency.getTimezone());
-
-    List<FareProduct> fareProducts = new ArrayList<>(dao.getAllFareProducts());
-    assertEquals(12, fareProducts.size());
-
-    FareProduct fp = fareProducts.stream().sorted(Comparator.comparing(FareProduct::getId)).findFirst().get();
-    assertEquals("id=31-day_disabled|category=disabled|medium=null", fp.getId().getId());
-    assertEquals("31-Day Pass Persons with Disabilities", fp.getName());
-    assertEquals("USD", fp.getCurrency());
-    assertEquals(15.0, fp.getAmount(), 0);
-    assertEquals(3, fp.getDurationUnit());
-    assertEquals(31, fp.getDurationAmount());
-    assertEquals(2, fp.getDurationType());
-    RiderCategory cat = fp.getRiderCategory();
-    assertEquals("Persons with Disabilities", cat.getName());
-    assertEquals("disabled", cat.getId().getId());
-
-
-    List<FareLegRule> fareLegRules = new ArrayList<>(dao.getAllFareLegRules());
-    assertEquals(12, fareLegRules.size());
-
-    FareLegRule flr = fareLegRules.stream().sorted(Comparator.comparing(FareLegRule::getId)).findFirst().get();
-    assertEquals("groupId=Turlock|product=31-day_disabled|network=null|fromArea=null|toArea=null", flr.getId());
-    assertEquals("Turlock", flr.getLegGroupId().getId());
-
-    List<RiderCategory> riderCats = new ArrayList<>(dao.getAllRiderCategories());
-    assertEquals(5, riderCats.size());
-
-    RiderCategory riderCat = riderCats.stream().sorted(Comparator.comparing(RiderCategory::getId)).filter(c -> c.getId().getId().equals("youth")).findAny().get();
-    assertEquals("youth", riderCat.getId().getId());
-    assertEquals("Youth Age 18 and Under", riderCat.getName());
-    assertEquals(18, riderCat.getMaxAge());
-    assertEquals(RiderCategory.MISSING_VALUE, riderCat.getMinAge());
-    assertEquals("http://www.turlocktransit.com/fares.html", riderCat.getEligibilityUrl());
-
-    assertTrue(dao.hasFaresV1());
-    assertTrue(dao.hasFaresV2());
-  }
-  @Test
-  public void mdotMetroFaresV2() throws CsvEntityIOException, IOException {
-    String agencyId = "1";
-    GtfsRelationalDao dao = processFeed(GtfsTestData.getMdotMetroFaresV2(),
-      agencyId, false);
-
-    Agency agency = dao.getAgencyForId(agencyId);
-    assertEquals(agencyId, agency.getId());
-    assertEquals("Maryland Transit Administration Metro Subway", agency.getName());
-
-    List<FareProduct> fareProducts = new ArrayList<>(dao.getAllFareProducts());
-    assertEquals(21, fareProducts.size());
-
-    FareProduct fp = fareProducts.stream().sorted(Comparator.comparing(FareProduct::getId)).findFirst().get();
-    assertEquals("id=core_local_1_day_fare|category=null|medium=charmcard", fp.getId().getId());
-    assertEquals("1-Day Pass - Core Service", fp.getName());
-    assertEquals("USD", fp.getCurrency());
-    assertEquals(4.6, fp.getAmount(), 0.01);
-
-    List<FareLegRule> fareLegRules = new ArrayList<>(dao.getAllFareLegRules());
-    assertEquals(7, fareLegRules.size());
-
-    FareLegRule flr = fareLegRules.stream().sorted(Comparator.comparing(FareLegRule::getId)).findFirst().get();
-    assertEquals("groupId=core_local_one_way_trip|product=core_local_1_day_fare|network=core|fromArea=null|toArea=null", flr.getId());
-    assertEquals("core_local_one_way_trip", flr.getLegGroupId().getId());
-
-    List<FareTransferRule> fareTransferRules = new ArrayList<>(dao.getAllFareTransferRules());
-    assertEquals(3, fareTransferRules.size());
-
-    FareTransferRule ftr = fareTransferRules.stream().sorted(Comparator.comparing(FareTransferRule::getId)).findFirst().get();
-    assertEquals("1_core_express_one_way_trip_1_core_express_one_way_trip_null_-999_5400", ftr.getId());
-    assertEquals(new AgencyAndId("1", "core_express_one_way_trip"), ftr.getFromLegGroupId());
-    assertEquals(-999, ftr.getTransferCount());
-    assertEquals(5400, ftr.getDurationLimit());
-
-    List<FareMedium> media = new ArrayList<>(dao.getAllFareMedia());
-    assertEquals(3, fareTransferRules.size());
-
-    FareMedium medium = media.stream().filter(c -> c.getId().getId().equals("charmcard_senior")).findFirst().get();
-    assertEquals("charmcard_senior", medium.getId().getId());
-    assertEquals("Senior CharmCard", medium.getName());
-
-    List<StopAreaElement> stopAreaElements = new ArrayList<>(dao.getAllStopAreas());
-    assertEquals(0, stopAreaElements.size());
-
-    List<Route> routes = new ArrayList<>(dao.getAllRoutes());
-    assertEquals(1, routes.size());
-    assertEquals("core", routes.get(0).getNetworkId());
-
-    assertFalse(dao.hasFaresV1());
-    assertTrue(dao.hasFaresV2());
-  }
-
-
-  @Test
-  public void testFaresV2Distance() throws IOException{
-    MockGtfs gtfs = MockGtfs.create();
-    gtfs.putMinimal();
-    gtfs.putLines("fare_products.txt", "fare_product_id, amount, currency", "" +
-            "fare_1,5,EUR");
-    gtfs.putLines("fare_leg_rules.txt", "network_id,min_distance,max_distance,distance_type,fare_product_id",
-            "bus,0,3,1,fare_1"
-    );
-    GtfsRelationalDao dao = processFeed(gtfs.getPath(), "1", false);
-    assertTrue(dao.getAllFareLegRules().stream().map(fareLegRule -> fareLegRule.getMaxDistance()).findFirst().get() == 3.0);
-    assertTrue(dao.getAllFareLegRules().stream().map(fareLegRule -> fareLegRule.getMinDistance()).findFirst().get() == 0.0);
-    assertTrue(dao.getAllFareLegRules().stream().map(fareLegRule -> fareLegRule.getDistanceType()).findFirst().get() == 1);
   }
 
   @Test
@@ -1043,27 +924,6 @@ public class GtfsReaderTest {
     Route route = reader.getEntityStore().getEntityForId(Route.class,
         new AgencyAndId("1", "R-10"));
     assertEquals("Ten, Ten", route.getLongName());
-  }
-
-  /****
-   * Private Methods
-   ****/
-
-  public static GtfsRelationalDao processFeed(File resourcePath, String agencyId,
-      boolean internStrings) throws IOException {
-
-    GtfsReader reader = new GtfsReader();
-    reader.setDefaultAgencyId(agencyId);
-    reader.setInternStrings(internStrings);
-
-    reader.setInputLocation(resourcePath);
-
-    GtfsRelationalDaoImpl entityStore = new GtfsRelationalDaoImpl();
-    entityStore.setGenerateIds(true);
-    reader.setEntityStore(entityStore);
-
-    reader.run();
-    return entityStore;
   }
 
   private ShapePoint getShapePoint(Iterable<ShapePoint> shapePoints,
