+++ onebusaway-gtfs/src/test/resources/org/onebusaway/gtfs/piercetransit-wa-us-stop-areas-flex/areas.txt
@@ -1,0 +1,3 @@
+area_id,area_name
+4210800,Spanaway Core
+4210813,JBLM Stops
