+++ onebusaway-gtfs/src/test/resources/org/onebusaway/gtfs/piercetransit-wa-us-stop-areas-flex/booking_rules.txt
@@ -1,0 +1,5 @@
+booking_rule_id,booking_type,prior_notice_duration_min,prior_notice_duration_max,prior_notice_start_day,prior_notice_start_time,prior_notice_last_day,prior_notice_last_time,prior_notice_service_id,message,pickup_message,drop_off_message,phone_number,info_url,booking_url
+booking_route_76310,0,,,,,,,,"The Ruston Runner provides on-demand transit service in Northwest Tacoma along Ruston Way. To request a ride, call 253-270-1340 or book a trip through the GOIN’ Rides for all mobile app.",,,253-270-1340,https://www.piercetransit.org/rustonrunner/,
+booking_route_76311,0,,,,,,,,"The Tideflats Runner provides on-demand transit service in the Port of Tacoma, Fife, and Edgewood. To request a ride, call 253-270-1340 or book a trip through the GOIN’ Rides for all mobile app.",,,253-270-1340,https://www.piercetransit.org/tideflatsrunner/,
+booking_route_76312,0,,,,,,,,"The Spanaway Runner provides on-demand transit service throughout the Spanaway, Midland, and Parkland areas. To request a ride, call 253-270-1340 or book a trip through the GOIN’ Rides for all mobile app.",,,253-270-1340,https://www.piercetransit.org/spanaway-parkland-midland-runner/,
+booking_route_76313,0,,,,,,,,"The JBLM Runner provides on-demand transit service to and from the Joint Base Lewis-McChord military base. To request a ride, call 253-377-4380. If traveling onto the base, riders must have an approved Department of Defense ID.",,,253-377-4380,https://www.piercetransit.org/jblmrunner/,
