+++ onebusaway-gtfs/src/test/resources/org/onebusaway/gtfs/piercetransit-wa-us-stop-areas-flex/calendar_attributes.txt
@@ -1,0 +1,3 @@
+service_id,service_description
+c_69336_b_80376_d_31,JBLM Runner (Weekday)
+c_69335_b_80375_d_127,Flex Year Round (All days of week)
