+++ onebusaway-gtfs/src/test/resources/org/onebusaway/gtfs/piercetransit-wa-us-stop-areas-flex/feed_info.txt
@@ -1,0 +1,2 @@
+feed_publisher_url,feed_publisher_name,feed_lang,feed_version,feed_license,feed_contact_email,feed_contact_url,feed_start_date,feed_end_date,feed_id
+http://www.trilliumtransit.com,"Trillium Solutions, Inc.",en,UTC: 28-Apr-2023 22:41,,support+piercetransit-wa-us@trilliumtransit.com,http://support.trilliumtransit.com,20230428,20231201,piercetransit-wa-us
