+++ onebusaway-gtfs/src/test/resources/org/onebusaway/gtfs/piercetransit-wa-us-stop-areas-flex/routes.txt
@@ -1,0 +1,5 @@
+agency_id,route_id,route_short_name,route_long_name,route_desc,route_type,route_url,route_color,route_text_color,route_sort_order,min_headway_minutes,eligibility_restricted,continuous_pickup,continuous_drop_off,tts_route_short_name,tts_route_long_name
+4971,76310,,Ruston Runner,,3,,398bc5,,0,,0,1,1,,
+4971,76311,,Tideflats Runner,,3,,3b6a9d,,1,,0,1,1,,
+4971,76312,,Spanaway Runner,,3,,00befd,000000,2,,0,1,1,,
+4971,76313,,JBLM Runner,,3,,00befd,000000,3,,0,1,1,,
