+++ onebusaway-gtfs/src/test/resources/org/onebusaway/gtfs/piercetransit-wa-us-stop-areas-flex/stop_areas.txt
@@ -1,0 +1,16 @@
+area_id,stop_id
+4210800,area_1074
+4210800,area_1075
+4210800,area_1076
+4210813,4210804
+4210813,4210805
+4210813,4210806
+4210813,4210807
+4210813,4210801
+4210813,4210809
+4210813,4210810
+4210813,4210811
+4210813,4210812
+4210813,4210808
+4210813,4210802
+4210813,4210803
