+++ onebusaway-gtfs/src/test/resources/org/onebusaway/gtfs/piercetransit-wa-us-stop-areas-flex/stop_times.txt
@@ -1,0 +1,15 @@
+trip_id,arrival_time,departure_time,stop_id,stop_sequence,stop_headsign,pickup_type,drop_off_type,shape_dist_traveled,timepoint,start_service_area_id,end_service_area_id,start_service_area_radius,end_service_area_radius,continuous_pickup,continuous_drop_off,pickup_booking_rule_id,drop_off_booking_rule_id,start_pickup_drop_off_window,end_pickup_drop_off_window,mean_duration_factor,mean_duration_offset,safe_duration_factor,safe_duration_offset,tts_stop_headsign,min_arrival_time,max_departure_time
+t_5582575_b_80375_tn_0,,,area_1072,1,,2,1,0,0,,,,,1,1,booking_route_76310,booking_route_76310,07:00:00,22:00:00,1,5.00,1,7.00,,,
+t_5582575_b_80375_tn_0,,,area_1072,2,,1,2,0,0,,,,,1,1,booking_route_76310,booking_route_76310,07:00:00,22:00:00,1,5.00,1,7.00,,,
+t_5586053_b_80375_tn_0,,,area_1073,1,,2,1,0,0,,,,,1,1,booking_route_76311,booking_route_76311,07:00:00,22:00:00,1,5.00,1,10.00,,,
+t_5586053_b_80375_tn_0,,,area_1073,2,,1,2,0,0,,,,,1,1,booking_route_76311,booking_route_76311,07:00:00,22:00:00,1,5.00,1,10.00,,,
+t_5586054_b_80375_tn_0,,,4210800,1,,2,1,0,0,,,,,1,1,booking_route_76312,booking_route_76312,07:00:00,22:00:00,1,5.00,1,10.00,,,
+t_5586054_b_80375_tn_0,,,4210800,2,,1,2,0,0,,,,,1,1,booking_route_76312,booking_route_76312,07:00:00,22:00:00,1,5.00,1,10.00,,,
+t_5586055_b_80375_tn_0,,,4210800,1,,2,1,0,0,,,,,1,1,booking_route_76312,booking_route_76312,07:00:00,22:00:00,1,5.00,1,10.00,,,
+t_5586055_b_80375_tn_0,,,area_1077,2,,1,2,0,0,,,,,1,1,booking_route_76312,booking_route_76312,07:00:00,22:00:00,1,5.00,1,10.00,,,
+t_5586056_b_80375_tn_0,,,area_1077,1,,2,1,0,0,,,,,1,1,booking_route_76312,booking_route_76312,07:00:00,22:00:00,1,5.00,1,10.00,,,
+t_5586056_b_80375_tn_0,,,4210800,2,,1,2,0,0,,,,,1,1,booking_route_76312,booking_route_76312,07:00:00,22:00:00,1,5.00,1,10.00,,,
+t_5586096_b_80376_tn_0,,,4210813,1,,2,1,0,0,,,,,1,1,booking_route_76313,booking_route_76313,07:00:00,09:30:00,1,5.00,1,7.00,,,
+t_5586096_b_80376_tn_0,,,4210813,2,,1,2,0,0,,,,,1,1,booking_route_76313,booking_route_76313,07:00:00,09:30:00,1,5.00,1,7.00,,,
+t_5586097_b_80376_tn_0,,,4210813,1,,2,1,0,0,,,,,1,1,booking_route_76313,booking_route_76313,15:00:00,18:30:00,1,5.00,1,7.00,,,
+t_5586097_b_80376_tn_0,,,4210813,2,,1,2,0,0,,,,,1,1,booking_route_76313,booking_route_76313,15:00:00,18:30:00,1,5.00,1,7.00,,,
