+++ onebusaway-gtfs/src/test/resources/org/onebusaway/gtfs/piercetransit-wa-us-stop-areas-flex/stops.txt
@@ -1,0 +1,19 @@
+stop_id,stop_code,platform_code,stop_name,stop_desc,stop_lat,stop_lon,zone_id,stop_url,location_type,parent_station,stop_timezone,position,direction,wheelchair_boarding,tts_stop_name
+4210574,,,Ruston Way,,47.2803611075431,-122.478135960265,,,0,,America/Los_Angeles,,,0,
+4210575,,,Port of Tacoma,,47.2530211408576,-122.375902952852,,,0,,America/Los_Angeles,,,0,
+4210576,,,"Spanaway, WA, USA",,47.1215482031301,-122.41115802604,,,0,,America/Los_Angeles,,,0,
+4210577,,,"Midland, WA, USA",,47.172908040322,-122.366108336615,,,0,,America/Los_Angeles,,,0,
+4210578,,,"Parkland, WA, USA",,47.1352304417448,-122.449156667491,,,0,,America/Los_Angeles,,,0,
+4210579,,,"Spanaway Buffer, WA, USA",,47.1406240812435,-122.383051288107,,,0,,America/Los_Angeles,,,0,
+4210801,,,Lakewood Transit Center,,47.1651533527433,-122.513206389954,,,0,,America/Los_Angeles,,,0,
+4210802,,,Lakewood Station (Bay 6),,47.1525830474275,-122.49918573403,,,0,,America/Los_Angeles,,,0,
+4210803,,,Bridgeport Way & Pacific Hwy SW (Northbound),,47.1508470880185,-122.50432023215,,,0,,America/Los_Angeles,,,0,
+4210804,,,Bridgeport Way & Pacific Hwy SW (Southbound),,47.1505105434807,-122.504607230921,,,0,,America/Los_Angeles,,,0,
+4210805,,,Bridgeport Way & San Francisco Ave SW (Southbound),,47.1416586202339,-122.503714676304,,,0,,America/Los_Angeles,,,0,
+4210806,,,Bridgeport Way & San Francisco Ave SW (Northbound),,47.1427796398538,-122.504098108707,,,0,,America/Los_Angeles,,,0,
+4210807,,,Visitor Main Gate - Fairway Rd,,47.1367431155588,-122.500633149635,,,0,,America/Los_Angeles,,,0,
+4210808,,,Barnes Blvd & D St SW,,47.1297618324364,-122.494075148601,,,0,,America/Los_Angeles,,,0,
+4210809,,,Olympic Dining Facility - Outer Dr,,47.1282959051506,-122.491026105129,,,0,,America/Los_Angeles,,,0,
+4210810,,,McChord Base Exchange (Barnes Blvd & 14th St),,47.1266199477849,-122.494126392815,,,0,,America/Los_Angeles,,,0,
+4210811,,,JBLM Commissary,,47.1254155072807,-122.49638058507,,,0,,America/Los_Angeles,,,0,
+4210812,,,Lincoln Blvd SW & B St SW,,47.1227158430177,-122.495208934161,,,0,,America/Los_Angeles,,,0,
